+++ src/data2notion_main.py
@@ -24,7 +24,7 @@ from data2notion.serialization import (
 logger = logging.getLogger("data2notion")
 
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 
 __plugin_api_version__ = 1.0